menu, service, ai update
