+++ Data/Script/halcyon/menu/single_deal_menu.lua
@@ -1,4 +1,4 @@
-require 'common'
+require 'origin.common'
 
 function SingleItemDealMenu()
     local SingleItemDealMenu = Class('MerchantMenu')
