--- Data/Script/halcyon/services/chatter/dialogue/default.xml
@@ -1,9 +1,0 @@
-<DialogueSet>
-  <Name>Default</Name>
-  <DexId/>
-
-  <!-- Lines for each topics -->
-  <Topics>
-
-  </Topics>
-</DialogueSet>
