--- Data/Script/halcyon/services/chatter/topics/topics.xml
@@ -1,14 +1,0 @@
-<!-- 
-    Definition file for topics that a talker entity
-    may bring up.
-    
-    Priority decides which goes first if a topic has a message available for "talking". Lower values are lower priorities. priority 0 and -1 are reserved.
--->
-<Topics>
-  <!-- Status -->
-  <Topic>
-    <Name>Status</Name>
-    <Priority>1</Priority>
-    <ScriptModule>services.chatter.topics.status</ScriptModule>
-  </Topic>  
-</Topics>
